+++ Firmware/MaD_Firmware/Librarys/RA8876/Graph.h
@@ -9,7 +9,7 @@ typedef struct graph_t
     double maxY;
     double minY;
     int x, y, w, h; // Position and size of graph area to copy
-    float markerY[MAX_GRAPH_MARKERS];
+    double markerY[MAX_GRAPH_MARKERS];
     int markerCount;
     char title[MAX_GRAPH_TITLE_LENGTH];
     char units[MAX_GRAPH_UNITS_LENGTH];
@@ -29,7 +29,7 @@ typedef struct graph_t
 } ModuleGraph;
 
 void module_set_graph(Module *window, ModuleGraph *graph, const char *title, const char *units);
-void module_graph_add_marker(Module *module, float value);
+void module_graph_add_marker(Module *module, double value);
 void module_graph_insert(Module *module, double value);
 void module_graph_set_range(Module *module, double maxY, double minY);
 int module_graph_get_max_data(Module *module);
